+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - GitHub Sponsors
+- feat: `get_nvimtree_glyphs` & `get_config` functions init
+- feat: tests with [plenary.nvim](https://github.com/nvim-lua/plenary.nvim)
 
 ### Changed